+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+- Add support for macOS, Linux, Windows, and any Unix-like OS that has a `tailscale` command-line client in the path.
+
 ## [0.1.0] - 2025-11-19
 
 Initial release